Add curses UI layer
curses layer implements a simple api to display a double-nested list as
a grid within the terminal. Calls to the function TerminalGrid.draw_grid
will draw the double-nested list to screen. Or, for control over color
of the cells within the grid, convert the grid to glyphs--using a
variant of grid_to_glyphs()--and modify color through glyphs' .attr
before finallying calling TerminalGrid.draw_glyphs()