add curses.pad wrapping primitive SwearJar
SwearJar is a container for wrapping curses pads. When the pad is moved
around, SwearJar intelligently resizes the pad to fit within the bounds
of the terminal. Normally, curses errors out when we attempt to draw
offscreen, but SwearJar makes the pad invisible wherever it would have
been drawn off-screen.